Add files via upload
This is the first working version of the Python script I wrote to scrape data from the Datawattch website about the production of the solar panels. The solar panels are rented and the system is a closed system (e.g. no data is written to and delivered by the Growatt server).